Update README.md
Epistle
=======

Since there's no code yet, a short essay on the name and motivation will 
have to do. 

Canadians will recognize the pun on "sin bin".  Kit winemaking is popular 
in Canada due to our having "some of the highest rates of taxes on 
cigarettes and alcohol in the world. These are sometimes referred to as 
sin taxes." [wikipedia]  However, one can brew wine or beer at home (not 
for resale) in fair quantity paying no alcohol taxes at all; or one can 
brew in an on-premises brew shop for home consumption provided you show up 
to bottle the finished product yourself (using a $2 plastic bottling wand 
from Ollivander's Tax Avoidance Wand Shop). This has encouraged a vibrant 
kit wine industry.  

Cellaring kit wine is not the same as cellaring commercial wine.  At home 
you're managing SO2 levels by dead reckoning. SO2 levels decline whenever 
the wine is handled and this differs from one brewer to another. Kits 
specify a "one size fits all" potassium metabisulphite treatment. For 
this and other reasons, the cellaring limit for kit wines ranges from 
one to four years, so there is little use of expensive natural long-cork.  

Short agglomerated corks (agglomerated=glue+corkdust) are widely used 
when wine is bottled for consumption within six months or less 
(within this group there is a devoted red-nose clientele for whom 
six months rarely exceeds six weeks).  

With cheap cork, one still has the hassle of storing the bottles flat, 
which exacerbates the susceptibility of the wine to daily or seasonal 
storage temperature variations.  (A dedicated wine fridge can consume 
$5 worth of electricity per bottle-year, and this doesn't even count 
that these appliances tend not to last more than five years. This is 
not something the average home brewer is going to do with bottles of 
wine that cost $4 to make.)  Most homemade wine ends up in a basement 
or a crawlspace.  Regardless of enclosure, when the temperature changes, 
the liquid volume changes, causing dramatic pressure variations in the tiny 
ullage volume.  With horizontal storage, this can force the wine itself 
past the cork and out of the bottle.  It also contributes to gas exchange 
with the surrounding environment bringing fresh oxygen into the bottle 
to replenish oxygen absorbed by the wine.  Oxygen is the sworn enemy 
of flavour and preservation.  

I prefer to age my kit wines to their optimal drinking age stored 
upright in boxes using a mixture of synthetic and Stelvin (screw cap) 
enclosures.  One fine day a screw cap will become available which 
matches or exceeds the oxygen exchange mechanics of natural cork, 
with less fussy storage demands (concerning humidity at least, if not 
also temperature).  While I await this glorious technology, I'm 
dividing each batch between enclosure types to determine for myself 
which method works best over different conditions and time.  

With many boxes of bottles crammed together into the most temperature 
consistent corner of the crawlspace, it's difficult to select the 
desired vintage without handling the bottles.  Mechanical vibration 
of wine is not encouraged by the druids of limestone crayeres.  
Voila! it occurred to me to print a QR code for the top of each bottle 
linking to a small web application giving me details on the bottle and 
vintage, tracking brew notes, and providing a way to input tasting notes 
on a per-bottle or per-batch basis.  

Voila^2 it occurred to me that this would be a good supplementary 
coding exercise as part of my coursework with MongoDB. 

SynVinQR references an upright kit wine storage system using QR codes 
and a smartphone to enable bottle selection without physically handling 
the bottles in storage.